Fix website monorepo cleanup regressions
