+++ website/apps/bittensor-website/package.json
@@ -7,18 +7,12 @@
     "build": "yarn build:docs-css && next build --webpack",
     "build:docs-css": "tailwindcss -i src/app/docs/docs.tw.css -o src/app/docs/docs.css --minify",
     "start": "next start",
-    "lint": "next lint",
+    "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
     "postinstall": "fumadocs-mdx"
   },
   "dependencies": {
-    "@polkadot/extension-dapp": "^0.58.4",
-    "@polkadot/extension-inject": "^0.58.3",
-    "@polkadot/util": "^13.4.3",
-    "@polkadot/util-crypto": "^13.4.3",
-    "@raofoundation/api": "*",
     "@raofoundation/ui": "*",
     "@svgr/webpack": "6.5.1",
-    "@tanstack/react-query": "^5.101.2",
     "@types/node": "^20.17.0",
     "@types/react": "^19.2.17",
     "@types/react-dom": "^19.2.3",
@@ -34,10 +28,8 @@
     "framer-motion": "10.12.16",
     "fumadocs-core": "16.11.1",
     "fumadocs-mdx": "15.1.0",
-    "gsap": "3.11.5",
     "lucide-react": "^1.23.0",
     "next": "16.2.10",
-    "next-mdx-remote": "4.4.1",
     "postcss-extend-rule": "4.0.0",
     "postcss-flexbugs-fixes": "5.0.2",
     "postcss-import": "15.1.0",
@@ -50,14 +42,12 @@
     "react-chartjs-2": "^5.2.0",
     "react-dom": "^19.2.7",
     "react-katex": "3.0.1",
-    "react-lottie": "1.2.3",
     "react-markdown": "8.0.7",
     "react-responsive": "^10.0.0",
     "react-slider": "^2.0.6",
     "react-spinners": "^0.15.0",
     "react-syntax-highlighter": "15.5.0",
     "react-tooltip": "^5.28.0",
-    "react-type-animation": "^3.2.0",
     "rehype-katex": "6.0.3",
     "rehype-katex-v7": "npm:rehype-katex@^7.0.1",
     "rehype-raw": "6.1.1",
@@ -74,7 +64,6 @@
     "@tailwindcss/cli": "^4.3.2",
     "@types/mdx": "^2.0.14",
     "@types/react-katex": "^3.0.4",
-    "@types/react-lottie": "1.2.6",
     "@types/react-syntax-highlighter": "15.5.6",
     "tsconfig": "*"
   }
