--- website/apps/bittensor-website/src/app/(pages-without-header)/Icons/AccountIcon.tsx
@@ -1,18 +1,0 @@
-import React from 'react';
-
-export const AccountIcon = ({width = 24, height = 24}) => {
-  return (
-    <svg
-      width={width}
-      height={height}
-      viewBox='0 0 18 16'
-      fill='none'
-      xmlns='http://www.w3.org/2000/svg'
-    >
-      <path
-        d='M13.7413 3.86209H16.2241C16.4436 3.86209 16.6541 3.94928 16.8093 4.10448C16.9645 4.25968 17.0516 4.47018 17.0516 4.68967V14.6207C17.0516 14.8401 16.9645 15.0506 16.8093 15.2058C16.6541 15.3611 16.4436 15.4482 16.2241 15.4482H1.32758C1.10809 15.4482 0.897595 15.3611 0.742393 15.2058C0.587191 15.0506 0.5 14.8401 0.5 14.6207V1.37934C0.5 1.15985 0.587191 0.949353 0.742393 0.794151C0.897595 0.638949 1.10809 0.551758 1.32758 0.551758H13.7413V3.86209ZM1.23571 4.5978V14.7125H16.3159V4.5978H1.23571ZM1.23571 1.28747V3.86209H13.0056V1.28747H1.23571ZM11.2586 9.24137H13.7413V9.65517V10.069H11.2586V9.65517V9.24137Z'
-        fill='currentColor'
-      />
-    </svg>
-  );
-};
