--- website/apps/bittensor-website/src/app/(pages-without-header)/Icons/ArrowBottomRight.tsx
@@ -1,18 +1,0 @@
-import React from 'react';
-
-export const ArrowBottomRight = ({width = 12, height = 12}) => {
-  return (
-    <svg
-      xmlns='http://www.w3.org/2000/svg'
-      width={width}
-      height={height}
-      viewBox='0 0 12 12'
-      fill='none'
-    >
-      <path
-        d='M4.65379 8.65133L4.66467 7.89529L7.36528 7.89529L3.0832 3.61321L3.61352 3.08289L7.8956 7.36497L7.8956 4.66436L8.65164 4.65348L8.65164 8.65133L4.65379 8.65133Z'
-        fill='currentColor'
-      />
-    </svg>
-  );
-};
