--- website/apps/bittensor-website/src/app/(pages-without-header)/Icons/ArrowDown.tsx
@@ -1,18 +1,0 @@
-import React from 'react';
-
-export const ArrowDown = ({width = 12, height = 12}) => {
-  return (
-    <svg
-      xmlns='http://www.w3.org/2000/svg'
-      width={width}
-      height={height}
-      viewBox='0 0 12 12'
-      fill='none'
-    >
-      <path
-        d='M3.17381 6.92257L3.71611 6.39566L5.62574 8.30529L5.62574 2.24951L6.37571 2.24951L6.37571 8.30529L8.28534 6.39566L8.82764 6.92257L6.00072 9.74949L3.17381 6.92257Z'
-        fill='currentColor'
-      />
-    </svg>
-  );
-};
