--- website/apps/bittensor-website/src/app/(pages-without-header)/Icons/ArrowTopLeft.tsx
@@ -1,18 +1,0 @@
-import React from 'react';
-
-export const ArrowTopLeft = ({width = 12, height = 12}) => {
-  return (
-    <svg
-      xmlns='http://www.w3.org/2000/svg'
-      width={width}
-      height={height}
-      viewBox='0 0 12 12'
-      fill='none'
-    >
-      <path
-        d='M7.34609 3.34818L7.33521 4.10423L4.63459 4.10423L8.91667 8.38631L8.38636 8.91662L4.10428 4.63454L4.10428 7.33516L3.34823 7.34604L3.34823 3.34818L7.34609 3.34818Z'
-        fill='currentColor'
-      />
-    </svg>
-  );
-};
