--- website/apps/bittensor-website/src/app/(pages-without-header)/Icons/ArrowUp.tsx
@@ -1,18 +1,0 @@
-import React from 'react';
-
-export const ArrowUp = ({width = 12, height = 12}) => {
-  return (
-    <svg
-      xmlns='http://www.w3.org/2000/svg'
-      width={width}
-      height={height}
-      viewBox='0 0 12 12'
-      fill='none'
-    >
-      <path
-        d='M8.82729 5.07645L8.28499 5.60336L6.37536 3.69374L6.37536 9.74951L5.62539 9.74951L5.62539 3.69374L3.71576 5.60336L3.17346 5.07645L6.00037 2.24954L8.82729 5.07645Z'
-        fill='currentColor'
-      />
-    </svg>
-  );
-};
