--- website/apps/bittensor-website/src/app/(pages-without-header)/Icons/ChevronUp.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-
-export const ChevronUp = ({width = 7, height = 5}) => {
-  return (
-    <svg
-      xmlns='http://www.w3.org/2000/svg'
-      width={width}
-      height={height}
-      viewBox='0 0 7 5'
-      fill='none'
-    >
-      <path d='M3.5 0.83789L6.9641 4.08113L0.035899 4.08113L3.5 0.83789Z' fill='currentColor' />
-    </svg>
-  );
-};
