--- website/apps/bittensor-website/src/app/(pages-without-header)/Icons/CubeIcon.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-export const CubeIcon = ({height = 16, className = ''}) => {
-  return (
-    <svg width='16' height='19' viewBox='0 0 16 19' fill='none' xmlns='http://www.w3.org/2000/svg'>
-      <path
-        d='M7.5 16.8434V9.41647L1 5.653V12.7338C1 12.8363 1.02564 12.9325 1.07692 13.0222C1.12821 13.112 1.20513 13.1889 1.3077 13.253L7.5 16.8434ZM8.5 16.8434L14.6923 13.253C14.7949 13.1889 14.8718 13.112 14.9231 13.0222C14.9744 12.9325 15 12.8363 15 12.7338V5.653L8.5 9.41647V16.8434ZM8 8.55685L14.425 4.84337L8.3077 1.303C8.20513 1.2389 8.10257 1.20685 8 1.20685C7.89743 1.20685 7.79487 1.2389 7.6923 1.303L1.575 4.84337L8 8.55685ZM0.807675 14.1376C0.555125 13.9928 0.357375 13.796 0.214425 13.5472C0.0714749 13.2985 0 13.0293 0 12.7395V5.5203C0 5.23057 0.0714749 4.96134 0.214425 4.71262C0.357375 4.46389 0.555125 4.26709 0.807675 4.12222L7.1923 0.443375C7.44487 0.298508 7.7141 0.226074 8 0.226074C8.2859 0.226074 8.55513 0.298508 8.8077 0.443375L15.1923 4.12222C15.4449 4.26709 15.6426 4.46389 15.7856 4.71262C15.9285 4.96134 16 5.23057 16 5.5203V12.7395C16 13.0293 15.9285 13.2985 15.7856 13.5472C15.6426 13.796 15.4449 13.9928 15.1923 14.1376L8.8077 17.8165C8.55513 17.9613 8.2859 18.0338 8 18.0338C7.7141 18.0338 7.44487 17.9613 7.1923 17.8165L0.807675 14.1376Z'
-        fill='currentColor'
-      />
-    </svg>
-  );
-};
