--- website/apps/bittensor-website/src/app/(pages-without-header)/Icons/SearchIcon.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-export const SearchIcon = ({height = 16, className = ''}) => {
-  return (
-    <svg width='18' height='18' viewBox='0 0 18 18' fill='none' xmlns='http://www.w3.org/2000/svg'>
-      <path
-        d='M16.4855 17.2838L10.224 11.0223C9.72399 11.4479 9.14899 11.7774 8.49899 12.0107C7.84899 12.2441 7.19578 12.3607 6.53937 12.3607C4.9377 12.3607 3.58216 11.8063 2.47274 10.6975C1.36331 9.58868 0.808594 8.23387 0.808594 6.63307C0.808594 5.03227 1.363 3.67642 2.47182 2.56552C3.58064 1.45462 4.93544 0.89917 6.53624 0.89917C8.13706 0.89917 9.49291 1.45388 10.6038 2.5633C11.7147 3.67271 12.2701 5.02826 12.2701 6.62995C12.2701 7.32481 12.1471 7.99725 11.9009 8.64725C11.6548 9.29725 11.3317 9.85302 10.9317 10.3146L17.1932 16.5761L16.4855 17.2838ZM6.53937 11.3607C7.8663 11.3607 8.98649 10.904 9.89994 9.99052C10.8134 9.07705 11.2701 7.95686 11.2701 6.62995C11.2701 5.30303 10.8134 4.18284 9.89994 3.26937C8.98649 2.3559 7.8663 1.89917 6.53937 1.89917C5.21245 1.89917 4.09226 2.3559 3.17879 3.26937C2.26534 4.18284 1.80862 5.30303 1.80862 6.62995C1.80862 7.95686 2.26534 9.07705 3.17879 9.99052C4.09226 10.904 5.21245 11.3607 6.53937 11.3607Z'
-        fill='currentColor'
-      />
-    </svg>
-  );
-};
