--- website/apps/bittensor-website/src/app/(pages-without-header)/Icons/StakeIcon.tsx
@@ -1,18 +1,0 @@
-import React from 'react';
-
-export const StakeIcon = ({width = 24, height = 24}) => {
-  return (
-    <svg
-      xmlns='http://www.w3.org/2000/svg'
-      width={width}
-      height={height}
-      viewBox='0 0 24 24'
-      fill='none'
-    >
-      <path
-        d='M11.4999 20.5002L11.4999 12.6021C10.5615 12.6021 9.66467 12.4268 8.80954 12.0762C7.95442 11.7255 7.19673 11.2201 6.53647 10.5598C5.87622 9.89958 5.37493 9.13805 5.03262 8.27525C4.69032 7.41243 4.51917 6.51307 4.51917 5.57715V4.57715H5.51917C6.43693 4.57715 7.32557 4.75625 8.18507 5.11445C9.04457 5.47266 9.80209 5.98074 10.4576 6.63867C10.9871 7.16816 11.4163 7.77008 11.7451 8.44445C12.074 9.1188 12.2916 9.83161 12.398 10.5829C12.5454 10.2996 12.7108 10.0332 12.8942 9.78387C13.0775 9.53451 13.2852 9.2938 13.5172 9.06175C14.1758 8.40381 14.9367 7.89574 15.8001 7.53752C16.6634 7.17932 17.5634 7.00022 18.4999 7.00022H19.4999V8.00022C19.4999 8.93717 19.3206 9.83752 18.962 10.7013C18.6034 11.565 18.0947 12.3263 17.4361 12.9851C16.7774 13.6439 16.0246 14.1445 15.1778 14.4868C14.331 14.8291 13.4384 15.0002 12.4999 15.0002L12.4999 20.5002H11.4999ZM11.5192 11.5771C11.5192 10.7771 11.365 10.0146 11.0567 9.28965C10.7483 8.56465 10.3108 7.91882 9.74417 7.35215C9.1775 6.78548 8.53167 6.34798 7.80667 6.03965C7.08167 5.73132 6.31917 5.57715 5.51917 5.57715C5.51917 6.37715 5.66917 7.14382 5.96917 7.87715C6.26917 8.61048 6.7025 9.26048 7.26917 9.82715C7.83583 10.3938 8.48583 10.8271 9.21917 11.1271C9.9525 11.4271 10.7192 11.5771 11.5192 11.5771ZM12.4999 14.0002C13.2999 14.0002 14.0624 13.8502 14.7874 13.5502C15.5124 13.2502 16.1583 12.8169 16.7249 12.2502C17.2916 11.6836 17.7291 11.0336 18.0374 10.3002C18.3458 9.56689 18.4999 8.80022 18.4999 8.00022C17.6999 8.00022 16.9333 8.15439 16.1999 8.46272C15.4666 8.77106 14.8166 9.20856 14.2499 9.77522C13.6833 10.3419 13.2499 10.9877 12.9499 11.7127C12.6499 12.4377 12.4999 13.2002 12.4999 14.0002Z'
-        fill='currentColor'
-      />
-    </svg>
-  );
-};
