--- website/apps/bittensor-website/src/app/(pages-without-header)/Icons/TransactionIcon.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-export const TransactionIcon = ({height = 16, className = ''}) => {
-  return (
-    <svg width='16' height='15' viewBox='0 0 16 15' fill='none' xmlns='http://www.w3.org/2000/svg'>
-      <path
-        d='M3.3077 14.1299L0 10.8222L3.3077 7.51451L4.02115 8.22218L1.92115 10.3222H8.21152V11.3222H1.92115L4.01537 13.4222L3.3077 14.1299ZM10.8087 11.4184C10.6401 11.4184 10.4984 11.3613 10.3837 11.2472C10.2689 11.1332 10.2115 10.9918 10.2115 10.8232C10.2115 10.6546 10.2686 10.5129 10.3826 10.3982C10.4967 10.2834 10.6381 10.226 10.8067 10.226C10.9753 10.226 11.117 10.2831 11.2317 10.3971C11.3465 10.5112 11.4038 10.6525 11.4038 10.8212C11.4038 10.9898 11.3468 11.1315 11.2327 11.2462C11.1187 11.361 10.9773 11.4184 10.8087 11.4184ZM14.001 11.4184C13.8324 11.4184 13.6907 11.3613 13.576 11.2472C13.4612 11.1332 13.4038 10.9918 13.4038 10.8232C13.4038 10.6546 13.4609 10.5129 13.575 10.3982C13.689 10.2834 13.8304 10.226 13.999 10.226C14.1676 10.226 14.3093 10.2831 14.424 10.3971C14.5388 10.5112 14.5962 10.6525 14.5962 10.8212C14.5962 10.9898 14.5391 11.1315 14.4251 11.2462C14.311 11.361 14.1696 11.4184 14.001 11.4184ZM12.6923 6.74526L11.9789 6.05681L14.0981 3.93758H7.78848V2.93756H14.0788L11.9846 0.837582L12.6923 0.129883L16 3.43758L12.6923 6.74526ZM2.001 4.03373C1.83238 4.03373 1.6907 3.9767 1.57595 3.86263C1.46122 3.74857 1.40385 3.60722 1.40385 3.43858C1.40385 3.26997 1.46088 3.12828 1.57495 3.01353C1.68902 2.89878 1.83037 2.84141 1.999 2.84141C2.16762 2.84141 2.3093 2.89845 2.42405 3.01253C2.53878 3.1266 2.59615 3.26794 2.59615 3.43656C2.59615 3.60519 2.53912 3.74687 2.42505 3.86161C2.31098 3.97636 2.16963 4.03373 2.001 4.03373ZM5.19333 4.03373C5.02471 4.03373 4.88302 3.9767 4.76827 3.86263C4.65352 3.74857 4.59615 3.60722 4.59615 3.43858C4.59615 3.26997 4.65318 3.12828 4.76725 3.01353C4.88133 2.89878 5.02268 2.84141 5.1913 2.84141C5.35992 2.84141 5.5016 2.89845 5.61635 3.01253C5.7311 3.1266 5.78848 3.26794 5.78848 3.43656C5.78848 3.60519 5.73144 3.74687 5.61738 3.86161C5.50329 3.97636 5.36194 4.03373 5.19333 4.03373Z'
-        fill='currentColor'
-      />
-    </svg>
-  );
-};
