--- website/apps/bittensor-website/src/app/components/Sections/Sections.module.css
@@ -1,142 +1,0 @@
-.section_1 {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: 0px 52px;
-  justify-content: flex-start;
-
-  &_text {
-    font-weight: 400;
-    max-width: 600px;
-    font-family: 'FiraCode';
-    font-size: 12px;
-    color: var(--text-primary);
-    line-height: 180%;
-    letter-spacing: 3%;
-    margin-bottom: 52px;
-    text-transform: uppercase;
-    animation-delay: 2s;
-    animation: slideIn 0.75s ease-out;
-  }
-}
-@keyframes slideIn {
-  from {
-    transform: translateX(-40px);
-    opacity: 0;
-  }
-  to {
-    transform: translateX(0px);
-    opacity: 1;
-  }
-}
-
-@keyframes slideUp {
-  from {
-    transform: translateY(40px);
-    opacity: 0;
-  }
-  to {
-    transform: translateY(0px);
-    opacity: 1;
-  }
-}
-
-/* Mobile*/
-@media (max-width: 700px) {
-  .section_1 {
-    justify-content: center;
-    height: calc(30vh - 52px - 52px - 20px);
-  }
-}
-
-.section_2 {
-  width: 100%;
-  background-color: #000000;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 575px;
-  &_div {
-    max-width: 500px;
-    display: flex;
-    justify-content: center;
-    max-height: 500px;
-  }
-}
-
-.section_3_div {
-  height: calc(100vh - 52px - 52px -20px);
-  margin: -100px auto auto auto;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  align-content: center;
-  flex-wrap: wrap;
-  gap: 85px;
-  padding: 20px 52px;
-}
-
-/* Mobile*/
-@media (max-width: 700px) {
-  .section_3_div {
-    flex-direction: column;
-    height: fit-content;
-    margin: auto;
-    gap: 85px;
-    padding: 52px;
-  }
-}
-
-.section_4 {
-  height: 291px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: #111111;
-}
-/* Mobile*/
-@media (max-width: 700px) {
-  .section_4 {
-    height: 575px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: #111111;
-  }
-}
-
-.section_5 {
-  display: none;
-  justify-content: center;
-  align-items: center;
-  margin: 0px 56px;
-
-  &_text {
-    max-width: 600px;
-    color: white;
-    letter-spacing: 3%;
-    line-height: 180%;
-    text-transform: capitalize;
-  }
-}
-
-/* Mobile*/
-@media (max-width: 700px) {
-  .section_5 {
-    display: flex;
-    background-color: black;
-    width: 100%;
-    margin: auto;
-    &_text {
-      padding: 52px;
-      line-height: 180%;
-      letter-spacing: 3%;
-      font-weight: 400;
-    }
-  }
-}
-
-.image {
-  width: 100%;
-  height: 100%;
-}
