--- website/apps/bittensor-website/src/app/components/Sections/api/getStats.ts
@@ -1,8 +1,0 @@
-export const getStats = async () => {
-  const response = await fetch('/api/stats');
-  const data = await response.json();
-
-  return {
-    totalStaked: data.total_stake || 0,
-  };
-};
