--- website/apps/bittensor-website/src/app/components/Wallet/WalletMedia.tsx
@@ -1,10 +1,0 @@
-'use client';
-
-export default function WalletMedia() {
-  return (
-    <img
-      src={'/images/wallet-new.webp'}
-      style={{objectFit: 'cover', width: '100%', height: '100%'}}
-    />
-  );
-}
