--- website/apps/bittensor-website/src/app/stores/useWalletVisibleStore.ts
@@ -1,17 +1,0 @@
-import {create} from 'zustand';
-
-export type UIStore = {
-  isVisible: boolean;
-  open: () => void;
-  close: () => void;
-};
-
-export const useWalletVisibleStore = create<UIStore>((set, get) => ({
-  isVisible: false,
-  open: () => {
-    set({isVisible: true});
-  },
-  close: () => {
-    set({isVisible: false});
-  },
-}));
