+++ website/packages/eslint-config-custom/index.js
@@ -1,5 +1,5 @@
 module.exports = {
-  extends: ["next", "turbo", "prettier"],
+  extends: ["next", "prettier"],
   rules: {
     "@next/next/no-html-link-for-pages": "off"
   },
