+++ website/packages/eslint-config-custom/package.json
@@ -7,7 +7,6 @@
     "eslint": "7.23.0",
     "eslint-config-next": "13.0.0",
     "eslint-config-prettier": "8.3.0",
-    "eslint-config-turbo": "latest",
     "eslint-plugin-react": "7.31.8"
   },
   "devDependencies": {
